cast int

+++ ocr.py
@@ -55,7 +55,7 @@ class OCR:
 
         img = img.crop((3, 0, img.size[0], img.size[1]))
         resized = img.resize(
-            (img.size[0] * size_factor, img.size[1] * size_factor), Image.ANTIALIAS
+            (int(img.size[0] * size_factor), int(img.size[1] * size_factor)), Image.ANTIALIAS
         )
         enhancer = ImageEnhance.Sharpness(resized)
         factor = 3
